Fix being unable to disable threshold
The threshold was defaulting to `relevance_count / 4` if the passed threshold value was not truthy, causing the method for disabling the threshold - passing zero, as noted in the documentation - to fail.

This patch changes the behaviour such the threshold is only calculated if the given value is null, which is what the value defaults to if it is not passed.  Any other value that is passed to the function will always be respected.