Update $this->primaryKey to $this->getKeyName()
$this->primaryKey doesn't work when the primary key is controlled by
another trait.  For example when using a trait to set the primary key to
'uuid' will cause SearchableTrait to assume the primaryKey is still
'id', inherited from Model.

Since this method was only introduced in Laravel 5.4 since the
refactoring of Eloquent, this will drop support for Laravel 4.  However,
I do not see this as a great concern as Laravel 4 is unsupported and so
is anything below Laravel 5.5.

Signed-off-by: r32rf <60923682+r32rf@users.noreply.github.com>